+++ R/simmer-class.R
@@ -23,10 +23,10 @@ Simmer <- R6Class("simmer",
           " | queue status: ", self$get_queue_count(name),
           "(", self$get_queue_size(name), ") }\n"
         ))
-      for (name in names(private$gen))
+      for (name in names(private$src))
         cat(paste0(
-          "{ Generator: ", name,
-          " | monitored: ", private$gen[[name]][["mon"]],
+          "{ Source: ", name,
+          " | monitored: ", private$src[[name]][["mon"]],
           " | n_generated: ", self$get_n_generated(name), " }\n"
         ))
       invisible(self)
@@ -52,13 +52,14 @@ Simmer <- R6Class("simmer",
       self
     },
 
-    run = function(until=1000) {
+    run = function(until=Inf) {
       run_(private$sim_obj, until)
       self
     },
 
     add_resource = function(name, capacity=1, queue_size=Inf, mon=TRUE, preemptive=FALSE,
-                            preempt_order=c("fifo", "lifo"), queue_size_strict=FALSE) {
+                            preempt_order=c("fifo", "lifo"), queue_size_strict=FALSE)
+    {
       check_args(
         name = "string",
         capacity = c("number", "schedule"),
@@ -97,7 +98,8 @@ Simmer <- R6Class("simmer",
     },
 
     add_generator = function(name_prefix, trajectory, distribution, mon=1,
-                             priority=0, preemptible=priority, restart=FALSE) {
+                             priority=0, preemptible=priority, restart=FALSE)
+    {
       check_args(
         name_prefix = "string",
         trajectory = "trajectory",
@@ -109,7 +111,62 @@ Simmer <- R6Class("simmer",
       )
       ret <- add_generator_(private$sim_obj, name_prefix, trajectory[],
                             make_resetable(distribution), mon, priority, preemptible, restart)
-      if (ret) private$gen[[name_prefix]] <- c(mon=mon)
+      if (ret) private$src[[name_prefix]] <- c(mon=mon)
+      self
+    },
+
+    add_dataframe = function(name_prefix, trajectory, data, mon=1, batch=50,
+                        col_time="time", time=c("interarrival", "absolute"),
+                        col_attributes=NULL, col_priority="priority",
+                        col_preemptible=col_priority, col_restart="restart")
+    {
+      check_args(
+        name_prefix = "string",
+        trajectory = "trajectory",
+        data = "data.frame",
+        mon = "flag",
+        batch = "number",
+        col_time = "string",
+        col_attributes = c("string vector", "NULL"),
+        col_priority = c("string", "NULL"),
+        col_preemptible = c("string", "NULL"),
+        col_restart = c("string", "NULL")
+      )
+      time <- match.arg(time)
+
+      col_attributes <- as.character(col_attributes)
+      col_priority <- intersect(col_priority, names(data))
+      col_preemptible <- intersect(col_preemptible, names(data))
+      col_restart <- intersect(col_restart, names(data))
+      col_names <- c(col_time, col_priority, col_preemptible, col_restart, col_attributes)
+      col_undef <- setdiff(col_names, names(data))
+
+      if (length(col_undef))
+        stop(get_caller(), ": columns '", paste0(col_undef, collapse="', '"),
+             "' are not defined in ", as.character(substitute(data)), call.=FALSE)
+
+      if (!length(col_attributes)) {
+        col_attributes <- setdiff(names(data), col_names)
+        col_names <- c(col_names, col_attributes)
+      }
+
+      for (col_name in col_names) {
+        if (!(is.numeric(data[[col_name]]) || is.logical(data[[col_name]])))
+          stop(get_caller(), ": column '", col_name, "' is not numeric", call.=FALSE)
+      }
+
+      if (any(is.na(data[[col_time]])) || any(data[[col_time]] < 0))
+        stop(get_caller(), ": time must be positive", call.=FALSE)
+
+      if (time == "absolute") {
+        if (is.unsorted(data[[col_time]]))
+          stop(get_caller(), ": unsorted absolute time provided", call.=FALSE)
+        data[[col_time]] <- c(data[[col_time]][1], diff(data[[col_time]]))
+      }
+
+      ret <- add_dataframe_(private$sim_obj, name_prefix, trajectory[], data, mon, batch,
+                            col_time, col_attributes, col_priority, col_preemptible, col_restart)
+      if (ret) private$src[[name_prefix]] <- c(mon=mon)
       self
     },
 
@@ -130,7 +187,7 @@ Simmer <- R6Class("simmer",
       monitor_data
     },
 
-    get_n_generated = function(generator) get_n_generated_(private$sim_obj, generator),
+    get_n_generated = function(source) get_n_generated_(private$sim_obj, source),
 
     get_name = function() get_name_(private$sim_obj),
 
@@ -179,13 +236,13 @@ Simmer <- R6Class("simmer",
     },
 
     # not exposed, internal use
-    get_generators = function() { private$gen },
+    get_sources = function() { private$src },
     get_resources = function() { private$res }
   ),
 
   private = list(
     sim_obj = NULL,
     res = list(),
-    gen = list()
+    src = list()
   )
 )